chore: add i18n + custom urls + svg as components + module resolver
I did a lot of work this weekend aiming to create a perfect i18n with the static export ability.

Right now, every page listed in next.config.js will be exported in the locales declared in the same file and the locale will be a pathname prefix with the corresponing locale shortname.

Also in Babel:

- svg as components: super useful for styling components
- module resolver: useful for brevety when importanting your own components
- transform-decorators-legacy: allow decorators usage